[Xiao]{feat}: java review

+++ src/objectLessons/example/exceptionHandling/ExceptionDemo.java
@@ -1,13 +1,13 @@
 package objectLessons.example.exceptionHandling;
 
 public class ExceptionDemo {
-  public static double divide(int argument) {
+  public static double divide(float argument) {
     return Math.floor(10 / argument);
   }
 
   public static void main(String[] args) {
     for (String x: args) {
-      int argument = Integer.parseInt(x);
+      float argument = Float.parseFloat(x);
       try {
         System.out.println("10 /" + argument + " = " + divide(argument));
       } catch (Exception e) {
